all: rename package to github.com/charlievieth/go-sqlite3
This project is a hard'ish fork of github.com/mattn/go-sqlite3 that will
not always be compatible with the original. Therefore I'm renaming it to
avoid confusion.

Regarding incompatibilities, the first is that mattn/go-sqlite3 allows
for the SQL passed to Query to contain multiple statements, but will
only execute the last one - this package will return an error if
Query is passed multiple SQL statements. The second incompatibility
is that this library will eventually remove the rarely used
RegisterAggregator since it uses reflect.MethodByName which disables
dead code removal. The goal is to keep the number of incompatibilities
to minimum, unless justified by correctness or performance reasons.

The overall goal of this package is to be more correct and performant
than mattn/go-sqlite3 while not being constrained by backwards
compatibility (apart from removing RegisterAggregator this package
should only introduce breaking changes to code that is already
incorrect).