ci: remove cifuzz, docker, and Windows tests
This commit temporarily removes the CIFuzz, Docker, and Windows tests.
It also simplifies the remaining Linux/macOS tests. I'm doing this since
I am iterating quite quickly on this branch and would like to speed up
the testing of PRs.

This commit also fixes the "on" trigger for the CI tests so that we
don't run duplicate tests.

TODO: Add back the Windows and CIFuzz, and re-enable Codecov.