always close db connection on error in Open
This commit fixes a bug where the db connection would not always be
closed when an error occurred in Open. Additionally, it makes sure that
we always unregister any callbacks associated with the connection, which
previously did not always happen. This change consolidates the error
handling logic (previously, it had to be done for each return statement)
which should make it more difficult to introduce this type of bug in the
future.