Take the declared column type into account when assigning string values
This commit changes nextSyncLocked to treat values that are stored as
TEXT in BLOB columns as BLOBs. Since this library uses the type of the
value passed to Exec/Query to determine the data type a Go string stored
in a column with the declared type of BLOB will have a sqlite3 data type
of TEXT. This change greatly improves performance when scanning a string
saved into a BLOB column into a byte slice or sql.RawBytes - since it
eliminates the need to convert the raw data to a string then back to a
byte slice.